+++ src/main/java/com/example/bank/entity/Account.java
@@ -3,6 +3,7 @@ package com.example.bank.entity;
 import com.fasterxml.jackson.annotation.JsonManagedReference;
 import jakarta.persistence.*;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -18,7 +19,7 @@ public class Account {
     private String holder;
 
     @Column(name = "balance", nullable = false)
-    private float balance;
+    private BigDecimal balance;
 
     @Column(name = "operations")
     @OneToMany(fetch = FetchType.LAZY, mappedBy = "id", targetEntity = Operation.class)
@@ -27,7 +28,7 @@ public class Account {
 
     public Account() {}
 
-    public Account(int id, String holder, float balance, List<Operation> operations) {
+    public Account(int id, String holder, BigDecimal balance, List<Operation> operations) {
         this.id = id;
         this.holder = holder;
         this.balance = balance;
@@ -50,11 +51,11 @@ public class Account {
         this.holder = holder;
     }
 
-    public float getBalance() {
+    public BigDecimal getBalance() {
         return balance;
     }
 
-    public void setBalance(float balance) {
+    public void setBalance(BigDecimal balance) {
         this.balance = balance;
     }
